# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Card Management API
  description: |
    REST API для управления банковскими картами, пользователями и переводами.
    
    ## Основные возможности:
    - Аутентификация пользователей с JWT токенами
    - Управление банковскими картами (создание, блокировка, активация, удаление)
    - Переводы между собственными картами пользователя
    - Система запросов на блокировку карт
    - История операций с картами
    - Управление пользователями (только для администраторов)
    
    ## Роли пользователей:
    - **ROLE_USER**: Обычный пользователь, может управлять своими картами и переводами
    - **ROLE_ADMIN**: Администратор, имеет полный доступ ко всем функциям
    
    ## Безопасность:
    - Все эндпоинты (кроме аутентификации) требуют JWT токен
    - Пароли хешируются с использованием BCrypt
    - Номера карт шифруются в базе данных
    - Пользователи могут работать только со своими данными
  version: 2.0.0
  contact:
    name: Bank REST API Support
    email: support@bankrest.com
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.bankrest.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/card-block-requests:
    get:
      tags:
        - Card Block Requests
      summary: Получить все запросы на блокировку (только для администраторов)
      operationId: getAllBlockRequests
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Список всех запросов на блокировку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockRequestPage'
        '401':
          description: Не авторизован
        '403':
          description: Доступ запрещен - требуется роль администратора

    post:
      tags:
        - Card Block Requests
      summary: Создать запрос на блокировку карты (только для пользователей)
      operationId: createBlockRequest
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlockRequestRequest'
      responses:
        '200':
          description: Запрос на блокировку успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockRequest'
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Не авторизован
        '403':
          description: Доступ запрещен - требуется роль пользователя
  /api/card-block-requests/my:
    get:
      tags:
        - Card Block Requests
      summary: Получить запросы на блокировку пользователя
      operationId: getMyBlockRequests
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Список запросов на блокировку пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockRequestPage'
        '401':
          description: Не авторизован
        '403':
          description: Доступ запрещен - требуется роль пользователя
  /api/card-block-requests/status/{status}:
    get:
      tags:
        - Card Block Requests
      summary: Получить запросы на блокировку по статусу (только для администраторов)
      operationId: getBlockRequestsByStatus
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: path
          required: true
          description: Статус запроса
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Список запросов на блокировку по статусу
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockRequestPage'
        '401':
          description: Не авторизован
        '403':
          description: Доступ запрещен - требуется роль администратора
  /api/card-block-requests/{requestId}/process:
    post:
      tags:
        - Card Block Requests
      summary: Обработать запрос на блокировку (только для администраторов)
      operationId: processBlockRequest
      security:
        - BearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          description: ID запроса на блокировку
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessBlockRequestRequest'
      responses:
        '200':
          description: Запрос на блокировку успешно обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBlockRequest'
        '400':
          description: Неверные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Не авторизован
        '403':
          description: Доступ запрещен - требуется роль администратора
        '404':
          description: Запрос на блокировку не найден
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
  schemas:
...
    Card:
      type: object
      required:
        - id
        - maskedNumber
        - owner
        - expiryDate
        - status
        - balance
        - user
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        maskedNumber:
          type: string
          description: Замаскированный номер карты
          example: "**** **** **** 1234"
        owner:
          type: string
          description: Имя владельца карты
          example: "John Doe"
        expiryDate:
          type: string
          format: date
          description: Дата истечения карты
          example: "2025-12-31"
        status:
          type: string
          enum: [ACTIVE, BLOCKED, EXPIRED]
          description: Статус карты
          example: "ACTIVE"
        balance:
          type: number
          format: decimal
          description: Баланс карты
          example: 1000.50
        user:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
          description: Дата создания карты
          example: "2024-01-15T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Дата последнего обновления карты
          example: "2024-01-15T10:30:00"
...
    User:
      type: object
      required:
        - id
        - username
        - roles
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: "user123"
        roles:
          type: array
          items:
            type: string
            enum: [ROLE_ADMIN, ROLE_USER]
          description: Роли пользователя
          example: ["ROLE_USER"]
        createdAt:
          type: string
          format: date-time
          description: Дата создания пользователя
          example: "2024-01-15T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Дата последнего обновления пользователя
          example: "2024-01-15T10:30:00"
...
    CardBlockRequest:
      type: object
      required:
        - id
        - card
        - requester
        - status
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: ID запроса на блокировку
          example: 1
        card:
          $ref: '#/components/schemas/Card'
        requester:
          $ref: '#/components/schemas/User'
        reason:
          type: string
          description: Причина блокировки
          example: "Потерял карту"
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
          description: Статус запроса
          example: "PENDING"
        admin:
          $ref: '#/components/schemas/User'
        adminComment:
          type: string
          description: Комментарий администратора
          example: "Запрос одобрен"
        createdAt:
          type: string
          format: date-time
          description: Дата создания запроса
          example: "2024-01-15T10:30:00"
        processedAt:
          type: string
          format: date-time
          description: Дата обработки запроса
          example: "2024-01-15T11:30:00"
    CardBlockRequestPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardBlockRequest'
        totalElements:
          type: integer
          description: Общее количество элементов
        totalPages:
          type: integer
          description: Общее количество страниц
        size:
          type: integer
          description: Размер страницы
        number:
          type: integer
          description: Номер текущей страницы
    CreateBlockRequestRequest:
      type: object
      required:
        - cardId
      properties:
        cardId:
          type: integer
          format: int64
          description: ID карты для блокировки
          example: 1
        reason:
          type: string
          description: Причина блокировки
          example: "Потерял карту"
    ProcessBlockRequestRequest:
      type: object
      required:
        - approved
      properties:
        approved:
          type: boolean
          description: Решение по запросу (true - одобрить, false - отклонить)
          example: true
        adminComment:
          type: string
          description: Комментарий администратора
          example: "Запрос одобрен"
...
    ValidationErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время ошибки
        status:
          type: integer
          description: HTTP код статуса
        error:
          type: string
          description: Тип ошибки
        details:
          type: object
          additionalProperties:
            type: string
          description: Детали ошибок валидации