# src/main/java/com/example/bankrest/dto/UserDto.java
package com.example.bankrest.dto;

import com.example.bankrest.entity.Role;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.Set;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UserDto {
    private Long id;
    private String username;
    private Set<Role> roles;
} 